Lowers, valids and int -Made Pieceno in Grid an int -Don't let people choose pieces that can't move anywhere -Changed a bunch of input based if statements to not be case sensitive

+++ Draughts/src/Grid.py
@@ -18,7 +18,7 @@ class Grid:
         self.setWidth(width)
         self.height = height
         self.setHeight(height)
-        self.pieceNo = (self.width/2) * self.rows
+        self.pieceNo = int((self.width/2) * self.rows)
         self.blackPiece = "b"
         self.whitePiece = "w"
         self.blackKing = "B"
